Math.Diophantine

Description

A module for solving quadratic diophantine equations.

Documentation

data Equation

A way to setup an equation in the form of:

 ax^2 + bxy + cy^2 + dx + ey + f = 0

Constructors

GeneralEquation Z Z Z Z Z Z
A general quadratic diophantine equation.

LinearEquation Z Z Z
dx + ey + f = 0

SimpleHyperbolicEquation Z Z Z Z
bxy + dx +ey + f = 0

ElipticalEquation Z Z Z Z Z Z
Eliptical equations.

ParabolicEquation Z Z Z Z Z Z
Parabolic equations.

HyperbolicEquation Z Z Z Z Z Z
Hyperbolic equations.

data Solution

The results of attempting to solve an Equation.

Constructors

ZxZ
All Integer pairs satisfy the equation.

NoSolutions
For all (x,y) in ZxZ

SolutionSet [(Z, Z)]
The set of pairs (x,y) that satisfy the equation. These are not in any particular order, and may contain duplicates.

type Z = Integer

An alias for Integer, used to shorten type signatures.

specializeEquation :: Equation -> Equation

Detirmines what kind of equation form a GeneralEquation fits. If you pass a non GeneralEquation to this function, it is the same as id.

solve :: Equation -> Solution

Determines what type of equation to solve for, and then calls the appropriate solve function. Example:

>>> ghci> solve (GeneralEquation 1 2 3 3 5 0)
[(-3,0),(-2,-1),(0,0),(1,-1)]

solveLinear :: Equation -> Solution

Solves for Equations in the form of dx + ey + f = 0

solveSimpleHyperbolic :: Equation -> Solution

Solves for Equations in the form of bxy + dx + ey + f = 0

solveEliptical :: Equation -> Solution

Solves for Equations in the form of ax^2 + bxy + cy^2 + dx + ey + f = 0 when b^2 - 4ac < 0

solveParabolic :: Equation -> Solution

Solves for Equations in the form of ax^2 + bxy + cy^2 + dx + ey + f = 0 when b^2 - 4ac = 0

toMaybeList :: Solution -> Maybe [(Z, Z)]

Extracts the list of solution pairs from a Solution.